refactor: extract discussion clinent's struct into model package
