+++ client/discussion/model/code_block.go
@@ -1,0 +1,7 @@
+package model
+
+type CodeBlock struct {
+	CodeId   int64                       `json:"codeId"`
+	Lines    []ExtractedLine             `json:"lines"`
+	Comments []DiscussionCommentResponse `json:"comments"`
+}
