+++ client/discussion/model/comment_scope_enum.go
@@ -1,0 +1,8 @@
+package model
+
+type CommentScopeEnum int
+
+const (
+	CommentScopeGlobal CommentScopeEnum = iota
+	CommentScopeLocal
+)
