+++ client/discussion/model/discussion_available_request.go
@@ -1,0 +1,6 @@
+package model
+
+type DiscussionAvailableRequest struct {
+	RepoId    int64 `json:"repoId"`
+	Available bool  `json:"available"`
+}
