+++ client/discussion/model/discussion_content_response.go
@@ -1,0 +1,8 @@
+package model
+
+type DiscussionContentResponse struct {
+	DiscussionId    int64                       `json:"discussionId"`
+	Contents        []FileContent               `json:"contents"`
+	GlobalComments  []DiscussionCommentResponse `json:"globalComments"`
+	GlobalReactions []DiscussionReaction        `json:"discussionReaction"`
+}
