+++ client/discussion/model/discussion_deadline.go
@@ -1,0 +1,7 @@
+package model
+
+import "time"
+
+type DiscussionDeadline struct {
+	Deadline *time.Time `json:"due_date"`
+}
