+++ client/discussion/model/discussion_list_response.go
@@ -1,0 +1,6 @@
+package model
+
+type DiscussionListResponse struct {
+	TotalCount  int64         `json:"totalCount"`
+	Discussions []*Discussion `json:"discussions"`
+}
