+++ client/discussion/model/extracted_line.go
@@ -1,0 +1,6 @@
+package model
+
+type ExtractedLine struct {
+	LineNumber int    `json:"lineNumber"`
+	Content    string `json:"content"`
+}
