+++ client/discussion/model/file_content.go
@@ -1,0 +1,6 @@
+package model
+
+type FileContent struct {
+	FilePath   string      `json:"filePath"`
+	CodeBlocks []CodeBlock `json:"codeBlocks"`
+}
