+++ client/discussion/model/update_assignee_request.go
@@ -1,0 +1,6 @@
+package model
+
+type UpdateAssigneeRequest struct {
+	DiscussionId int64 `json:"discussionId"`
+	AssigneeId   int64 `json:"assigneeId"`
+}
